ListAdapter (и DiffUtil) - Изменено. Используется. Так же было бы более логично переименовать typealias - Изменено. Метод formatNumber можно вынести в отдельный файл чтобы избежать дублирования кода. - Вынес в пакет util.
