+++ app/src/main/java/ru/netology/nmedia/activity/MainActivity.kt
@@ -13,10 +13,6 @@ import java.text.DecimalFormat
 import kotlin.math.pow
 
 class MainActivity : AppCompatActivity() {
-    //var _binding: ActivityMainBinding? = null
-    //val binding: ActivityMainBinding
-        //get() = _binding!!
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         val binding = ActivityMainBinding.inflate(layoutInflater)
@@ -32,34 +28,9 @@ class MainActivity : AppCompatActivity() {
         })
         binding.list.adapter = adapter
         viewModel.data.observe(this) { posts ->
-            adapter.list = posts
+            adapter.submitList(posts)
         }
 
     }
-
-    /*binding.like.setOnClickListener {
-        viewModel.like()
-    }
-
-    binding.share.setOnClickListener {
-        viewModel.share()
-    }
-
-    binding.views.setOnClickListener {
-        viewModel.views()
-    }
-
-} }*/
-private fun formatNumber(number: Int): String {
-        if (number < 1000) {
-            return number.toString()
-        }
-
-        val suffixes = arrayOf("", "K", "M")
-        val formatter = DecimalFormat("#,##0.#")
-        val exp = (Math.log10(number.toDouble()) / 3).toInt()
-        return formatter.format(number / 10.0.pow(exp * 3)) + suffixes[exp]
-    }
-
 }
