+++ app/src/main/java/ru/netology/nmedia/adapter/PostAdapter.kt
@@ -1,24 +1,27 @@
 package ru.netology.nmedia.adapter
 
 
+import ru.netology.nmedia.utils.FormatNumber.formatNumber
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import ru.netology.nmedia.R
 import ru.netology.nmedia.databinding.CardPostBinding
 import ru.netology.nmedia.dto.Post
-import java.text.DecimalFormat
-import kotlin.math.pow
 
 typealias OnLikeListener = (post: Post) -> Unit
+typealias OnShareListener = (post: Post) -> Unit
+typealias OnViewListener = (post: Post) -> Unit
 
-class PostsAdapter(private val onLikeListener: OnLikeListener, private val onShareListener: OnLikeListener, private val onViewListener: OnLikeListener) :
-    RecyclerView.Adapter<PostViewHolder>() {
-    var list = emptyList<Post>()
+class PostsAdapter(private val onLikeListener: OnLikeListener, private val onShareListener: OnShareListener, private val onViewListener: OnViewListener) :
+    ListAdapter<Post, PostViewHolder>(PostDiffCallback()) {
+    /*var list = emptyList<Post>()
         set(value) {
             field = value
             notifyDataSetChanged()
-        }
+        }*/
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): PostViewHolder {
         val binding = CardPostBinding.inflate(LayoutInflater.from(parent.context), parent, false)
@@ -26,19 +29,19 @@ class PostsAdapter(private val onLikeListener: OnLikeListener, private val onSha
     }
 
     override fun onBindViewHolder(holder: PostViewHolder, position: Int) {
-        val post = list[position]
+        val post = getItem(position)
         holder.bind(post)
     }
 
-    override fun getItemCount(): Int = list.size
+    //override fun getItemCount(): Int = list.size
 
 }
 
 class PostViewHolder(
     private val binding: CardPostBinding,
     private val onLikeListener: OnLikeListener,
-    private val onShareListener: OnLikeListener,
-    private val onViewListener: OnLikeListener
+    private val onShareListener: OnShareListener,
+    private val onViewListener: OnViewListener
 ) : RecyclerView.ViewHolder(binding.root) {
     fun bind(post: Post) {
         binding.apply {
@@ -64,13 +67,14 @@ class PostViewHolder(
             }
         }
     }
-    private fun formatNumber(number: Int): String {
-        if (number < 1000) {
-            return number.toString()
-        }
-        val suffixes = arrayOf("", "K", "M")
-        val formatter = DecimalFormat("#,##0.#")
-        val exp = (Math.log10(number.toDouble()) / 3).toInt()
-        return formatter.format(number / 10.0.pow(exp * 3)) + suffixes[exp]
+}
+class PostDiffCallback : DiffUtil.ItemCallback<Post>() {
+    override fun areItemsTheSame(oldItem: Post, newItem: Post): Boolean {
+        return oldItem.id == newItem.id
     }
+
+    override fun areContentsTheSame(oldItem: Post, newItem: Post): Boolean {
+        return oldItem == newItem
+    }
+
 }
