+++ app/src/main/java/ru/netology/nmedia/dto/Post.kt
@@ -6,7 +6,7 @@ data class Post(
     val content: String,
     val published: String,
     var likes: Int = 0,
-    var shareds: Int = 0,
+    var shareds: Int = 1000,
     var viewers: Int = 0,
     var likedByMe: Boolean = false
 )
