+++ app/src/main/java/ru/netology/nmedia/viewmodel/PostViewModel.kt
@@ -8,7 +8,7 @@ import ru.netology.nmedia.repository.PostRepositoryInMemoryImpl
 
 class PostViewModel: ViewModel() {
     private val repository: PostRepository = PostRepositoryInMemoryImpl()
-    val data: LiveData<List<Post>> = repository.getAll()
+    val data = repository.getAll()
     fun likeById(id: Long) = repository.likeById(id)
     fun shareById(id: Long) = repository.shareById(id)
     fun viewsById(id: Long) = repository.viewsById(id)
